# docs/api/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DocReader PoC API",
    "description": "Local proof of concept for document ingestion, OCR, classification and extraction.\n\n**There is no authentication.** Every endpoint is anonymous, the Swagger page is public and\nno bearer token is required. Do not expose this API to the internet and do not send real\npersonal documents to it. Use synthetic, masked or authorized files only.\n\nConventions: JSON in camelCase, timestamps in UTC ISO 8601, errors as\n`application/problem+json`, and an `X-Correlation-Id` header on every response.\n\nStage 3 of the execution plan delivers upload, listing, detail, content, deletion, OCR\n(PP-OCRv5 on CPU, see ADR 0002), the raw text and the structured result of seven document\ntypes: CPF card, CIN/RG, CNH, proof of address, CNPJ card, CCMEI and social contract.\nOther document types are read as raw text only.",
    "version": "v1"
  },
  "paths": {
    "/api/v1/documents": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Receives a document and queues it for processing.",
        "description": "The response is asynchronous on purpose: the API stores the original, issues the protocol and\nanswers `202 Accepted` with a `Location` header, while OCR, classification and\nextraction run in the worker. Poll `statusUrl` to follow the document.\n            \nAccepted formats are PDF, PNG, JPEG and TIFF, checked by the real file signature rather than\nby the announced content type. Size and page limits are configurable and answer 413 and 422.\n            \nSend the optional `Idempotency-Key` header to make a retry safe: the same key with the\nsame file replays the original response, and the same key with a different file answers 409.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional replay protection key.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Upload-Channel",
            "in": "header",
            "description": "Origin of the upload, `WEB` or `API`. The interface sends WEB; anything else\ndefaults to API.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "description": "The document itself. PDF, PNG, JPEG or TIFF, validated by its real signature and not by the\nannounced content type.",
                    "format": "binary"
                  },
                  "ExpectedDocumentType": {
                    "maxLength": 64,
                    "type": "string",
                    "description": "Optional hint about the document type. It never overrides the classifier; a divergence is\nrecorded instead."
                  },
                  "ExternalReference": {
                    "maxLength": 128,
                    "type": "string",
                    "description": "Optional reference of the calling system, echoed back in every response."
                  },
                  "ProductServiceCode": {
                    "maxLength": 64,
                    "type": "string",
                    "description": "Code of the product or service the document is for. Unknown or inactive codes answer 422."
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "ExpectedDocumentType": {
                  "style": "form"
                },
                "ExternalReference": {
                  "style": "form"
                },
                "ProductServiceCode": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/documents",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "202": {
            "description": "Document accepted, stored and queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed or a field is invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/documents",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "409": {
            "description": "The Idempotency-Key was already used with a different file.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/conflict",
                  "title": "Idempotency-Key conflict",
                  "status": 409,
                  "detail": "This Idempotency-Key was already used with a different file. Use a new key or resend the original file.",
                  "instance": "/api/v1/documents",
                  "errorCode": "IDEMPOTENCY_KEY_CONFLICT",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "413": {
            "description": "The file is above the configured size limit.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/file-too-large",
                  "title": "File too large",
                  "status": 413,
                  "detail": "The uploaded file has 41943040 bytes and the limit is 26214400 bytes.",
                  "instance": "/api/v1/documents",
                  "errorCode": "FILE_TOO_LARGE",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "415": {
            "description": "The real signature is not an accepted format, or contradicts the declared type.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unsupported-media-type",
                  "title": "Unsupported media type",
                  "status": 415,
                  "detail": "Only PDF, PNG, JPEG and TIFF files are accepted.",
                  "instance": "/api/v1/documents",
                  "errorCode": "UNSUPPORTED_MEDIA_TYPE",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "422": {
            "description": "The file is an accepted format but unreadable or above the page limit, or `productServiceCode` is unknown or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unprocessable-content",
                  "title": "Content cannot be processed",
                  "status": 422,
                  "detail": "The document has 84 pages and the limit is 50 pages.",
                  "instance": "/api/v1/documents",
                  "errorCode": "PAGE_LIMIT_EXCEEDED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Lists uploaded documents, newest first.",
        "description": "Every filter is optional and they combine with AND. `page` starts at 1 and\n`pageSize` is capped by the configured maximum. The envelope carries\n`totalCount` and `totalPages` so a client can paginate without guessing.",
        "parameters": [
          {
            "name": "protocol",
            "in": "query",
            "description": "Full or partial protocol, case insensitive.",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          },
          {
            "name": "fileName",
            "in": "query",
            "description": "Full or partial original file name, case insensitive.",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "documentType",
            "in": "query",
            "description": "Detected or expected document type, matched exactly.",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "description": "Upload channel: WEB or API.",
            "schema": {
              "maxLength": 16,
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status: RECEIVED, STORED, QUEUED, PREPROCESSING, OCR_RUNNING, CLASSIFYING, EXTRACTING,\nCOMPLETED, FAILED or REJECTED.",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          },
          {
            "name": "productServiceCode",
            "in": "query",
            "description": "Code of the product or service the document is linked to, case insensitive.",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "externalReference",
            "in": "query",
            "description": "Full or partial external reference sent at upload time, case insensitive.",
            "schema": {
              "maxLength": 256,
              "type": "string"
            }
          },
          {
            "name": "uploadedFrom",
            "in": "query",
            "description": "Lower bound of the upload instant, inclusive, ISO 8601 in UTC.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "uploadedTo",
            "in": "query",
            "description": "Upper bound of the upload instant, inclusive, ISO 8601 in UTC.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "One based page number. Defaults to 1.",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Items per page. Defaults to the configured page size and is capped by the maximum.",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/documents",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "Page of documents, possibly empty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummaryResponsePagedResponse"
                }
              }
            }
          },
          "400": {
            "description": "A filter value is invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/documents",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentLinks": {
        "type": "object",
        "properties": {
          "self": {
            "type": "string",
            "description": "Consolidated detail."
          },
          "status": {
            "type": "string",
            "description": "Lightweight status for polling."
          },
          "content": {
            "type": "string",
            "description": "Original file, inline by default."
          },
          "download": {
            "type": "string",
            "description": "Original file as an attachment."
          },
          "text": {
            "type": "string",
            "description": "Raw OCR text, page by page."
          },
          "result": {
            "type": "string",
            "description": "Canonical structured result."
          },
          "reprocess": {
            "type": "string",
            "description": "POST here to run the pipeline again."
          }
        },
        "additionalProperties": false,
        "description": "Endpoints related to one document, so a client never has to build paths by hand."
      },
...
      "DocumentStatus": {
        "enum": [
          "RECEIVED",
          "STORED",
          "QUEUED",
          "PREPROCESSING",
          "OCR_RUNNING",
          "CLASSIFYING",
          "EXTRACTING",
          "COMPLETED",
          "FAILED",
          "REJECTED",
          "PURGED"
        ],
        "type": "string",
        "description": "Lifecycle of a document, as defined by RF-007 of the PRD."
      },
...
      "DocumentSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity of the document.",
            "format": "uuid"
          },
          "protocol": {
            "type": "string",
            "description": "Human readable protocol."
          },
          "fileName": {
            "type": "string",
            "description": "Original file name as sent by the client."
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type detected from the file signature."
          },
          "sizeBytes": {
            "type": "integer",
            "description": "Size in bytes.",
            "format": "int64"
          },
          "pageCount": {
            "type": "integer",
            "description": "Number of pages.",
            "format": "int32"
          },
          "channel": {
            "$ref": "#/components/schemas/UploadChannel"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "expectedDocumentType": {
            "type": "string",
            "description": "Type hint sent at upload time, when any.",
            "nullable": true
          },
          "detectedDocumentType": {
            "type": "string",
            "description": "Type identified by the classifier, null until the document is classified.",
            "nullable": true
          },
          "classificationConfidence": {
            "type": "number",
            "description": "Confidence between 0 and 1, when available.",
            "format": "double",
            "nullable": true
          },
          "externalReference": {
            "type": "string",
            "description": "Caller reference sent at upload time, when any.",
            "nullable": true
          },
          "productService": {
            "$ref": "#/components/schemas/ProductServiceReferenceResponse"
          },
          "expiresAt": {
            "type": "string",
            "description": "When the document becomes eligible for purge, in UTC; null when no retention was applied.",
            "format": "date-time",
            "nullable": true
          },
          "uploadedAt": {
            "type": "string",
            "description": "Upload instant in UTC.",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "description": "Instant processing finished, in UTC.",
            "format": "date-time",
            "nullable": true
          },
          "links": {
            "$ref": "#/components/schemas/DocumentLinks"
          }
        },
        "additionalProperties": false,
        "description": "One row of the listing.",
        "example": {
          "id": "0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21",
          "protocol": "DOC-20260924-000001",
          "fileName": "cartao-cnpj.pdf",
          "mimeType": "application/pdf",
          "sizeBytes": 184320,
          "pageCount": 1,
          "channel": "API",
          "status": "QUEUED",
          "expectedDocumentType": null,
          "detectedDocumentType": null,
          "classificationConfidence": null,
          "externalReference": "CLIENTE-123",
          "productService": {
            "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
            "code": "CONTA-PJ",
            "name": "Abertura de conta PJ"
          },
          "expiresAt": "2027-09-24T22:00:00Z",
          "uploadedAt": "2026-09-24T22:00:00Z",
          "completedAt": null,
          "links": {
            "self": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21",
            "status": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/status",
            "content": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/content",
            "download": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/content?download=true",
            "text": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/text",
            "result": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/result",
            "reprocess": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/reprocess"
          }
        }
      },
      "DocumentSummaryResponsePagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentSummaryResponse"
            },
            "description": "Items of the current page, newest first."
          },
          "page": {
            "type": "integer",
            "description": "One based page number.",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Number of items per page.",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Total number of items matching the filters.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "Number of pages for the current page size.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Envelope of every paginated listing.",
        "example": {
          "items": [
            {
              "id": "0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21",
              "protocol": "DOC-20260924-000001",
              "fileName": "cartao-cnpj.pdf",
              "mimeType": "application/pdf",
              "sizeBytes": 184320,
              "pageCount": 1,
              "channel": "API",
              "status": "QUEUED",
              "expectedDocumentType": null,
              "detectedDocumentType": null,
              "classificationConfidence": null,
              "externalReference": "CLIENTE-123",
              "productService": {
                "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
                "code": "CONTA-PJ",
                "name": "Abertura de conta PJ"
              },
              "expiresAt": "2027-09-24T22:00:00Z",
              "uploadedAt": "2026-09-24T22:00:00Z",
              "completedAt": null,
              "links": {
                "self": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21",
                "status": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/status",
                "content": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/content",
                "download": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/content?download=true",
                "text": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/text",
                "result": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/result",
                "reprocess": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/reprocess"
              }
            }
          ],
          "page": 1,
          "pageSize": 20,
          "totalCount": 1,
          "totalPages": 1
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProductServiceReferenceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Unique code."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        },
        "additionalProperties": false,
        "description": "The product or service a document is linked to."
      },
...
      "UploadAcceptedResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity of the document.",
            "format": "uuid"
          },
          "protocol": {
            "type": "string",
            "description": "Human readable protocol, format DOC-yyyyMMdd-NNNNNN."
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "statusUrl": {
            "type": "string",
            "description": "Lightweight endpoint for polling."
          },
          "documentUrl": {
            "type": "string",
            "description": "Consolidated detail endpoint."
          },
          "contentUrl": {
            "type": "string",
            "description": "Endpoint that serves the original file."
          }
        },
        "additionalProperties": false,
        "description": "Body of the 202 answered by the upload endpoint. The document is persisted and queued; the OCR\nruns afterwards, outside the request.",
        "example": {
          "id": "0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21",
          "protocol": "DOC-20260924-000001",
          "status": "QUEUED",
          "statusUrl": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/status",
          "documentUrl": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21",
          "contentUrl": "/api/v1/documents/0199c1f0-7b3a-7a10-9c44-2f1d8e6b4a21/content"
        }
      },
      "UploadChannel": {
        "enum": [
          "API",
          "WEB"
        ],
        "type": "string",
        "description": "How the document entered the system."
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}